Add `push-new-tag` and `retires`
